requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.6
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pygame==2.6.1
pyparsing==3.2.1
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
six==1.17.0
